+++ Server/package.json
@@ -4,7 +4,7 @@
   "description": "fast",
   "main": "app.js",
   "scripts": {
-    "test": "mocha test/**/*.js",
+    "test": "mocha --compilers js:babel-core/register test/**/*.js",
     "start": "nodemon app.js --exec babel-node --"
   },
   "repository": {
@@ -18,6 +18,12 @@
   },
   "homepage": "https://github.com/darothub/quikfood#readme",
   "dependencies": {
+    "babel": {
+      "presets": [
+        "stage-3",
+        "latest"
+      ]
+    },
     "babel-cli": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
     "body-parser": "^1.18.3",
