+++ Server/test/test.spec.js
@@ -4,6 +4,7 @@ import chaiHttp from 'chai-http';
 
 import server from '../app';
 
+
 chai.should();
 chai.use(chaiHttp);
 
@@ -25,16 +26,27 @@ describe('GET /api/v1/orders', () => {
       .get('/api/v1/orders')
       .end((err, res) => {
         res.should.have.status(200);
-        res.should.be.json();
-        res.body.should.be.a('array');
-        res.body[0].should.have.property('id');
-        res.body[0].should.have.property('name');
-        res.body[0].should.have.property('designation');
-        res.body[0].should.have.property('dishType');
-        res.body[0].should.have.property('drink');
-        res.body[0].should.have.property('qty');
-        res.body[0].should.have.property('price');
-        res.body[0].should.have.property('status');
+        res.body.should.have.property('result');
+        done();
+      });
+  });
+});
+
+describe('GET /api/v1/orders/:id', () => {
+  it('should return error if not found', (done) => {
+    chai.request(server)
+      .get('/api/v1/orders/:id')
+      .end((err, res) => {
+        res.should.have.status(404);
+        done();
+      });
+  });
+  it('should return a single order', (done) => {
+    chai.request(server)
+      .get('/api/v1/orders/:id')
+      .end((err, res) => {
+        res.should.have.status(200);
+        res.body.should.be.a('object');
         done();
       });
   });
